Day 1 - Two Colour LED Module
Instead of the supplied 220R resistors, I found I got better results
putting a 680R on the red pin, and a 100R on the red pin.

This gave an even brightness between red and green, and a
distinguishable orange/yellow colour.

With the 220R's, the red completely swamped the green. Red and green
together was barely different to red only.